+++ Konsole/IConsole.cs
@@ -3,6 +3,7 @@ using Konsole.Internal;
 
 namespace Konsole
 {
+
     public interface IConsole : IWriteLine
     {
         ConsoleState State { get; set; }
@@ -10,6 +11,8 @@ namespace Konsole
         int WindowHeight { get; }
         int CursorTop { get; set; }
         int CursorLeft { get; set; }
+        Colors Colors { get; set; }
+        void DoCommand(IConsole console, Action action);
         ConsoleColor ForegroundColor { get; set; }
         ConsoleColor BackgroundColor { get; set; }
         void PrintAt(int x, int y, string format, params object[] args);
@@ -17,6 +20,9 @@ namespace Konsole
         void PrintAt(int x, int y, char c);
         void ScrollUp();
         void Clear();
+
+        // saves the current cursor position, (left and top) as well as colors
+        // runs the command, then restores the state.
 
         void MoveBufferArea(int sourceLeft, int sourceTop, int sourceWidth, int sourceHeight, int targetLeft,
             int targetTop, char sourceChar, ConsoleColor sourceForeColor, ConsoleColor sourceBackColor);