+++ change-log.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 New core `Window` functionality, plus big class renames, main interface `IConsole` has changed from `1.0` and `2.0`.
 
 ### Breaking changes
+- rename `Form.Show` to `Form.Write` (form is written at current cursor position, and position is updated)
 - change `IConsole` `WindowWidth` from a `method` to a `property`.
 - rename `Console` to `Window`
 - rename `ConsoleWriter` to `Writer`
@@ -27,6 +28,7 @@ New core `Window` functionality, plus big class renames, main interface `IConsol
   * `.BufferHilighted` `string[]` returns an 'approve-able' text buffer where each character is represented by 2 characters with one of of them representing the background color of the buffer.
 
 ### Backlog (not all will go into 2.0)
+- appharbor CI/CD testing with fake or cake.
 - rename 'echo' to (isMockConsole and invert) or, override behavior and remove altogether from Window.
 - investigate supporting clipping when printing i.e. overlapping windows - this may not be possible without being very slow, or using native window calls. This might also contradict the nature of the 'window' as being a thing that allows you to quickly print to your window, so a sub window is just a shortcut to actually print to that window.
 - (done) be more consistent, move IConsole to first parameter